// src/modules/unit/ts/custom-status.enum.ts
export enum CustomStatus {
  CancelPVZ = 'Отмена ПВЗ',
  InstantCancel = 'Моментальная отмена',
  AwaitingDelivery = 'Ожидает доставки',
  AwaitingPackaging = 'Ожидает сборки',
  Delivering = 'Доставляется',
  Delivered = 'Доставлен',
  Return = 'Возврат',
  AwaitingPayment = 'Ожидаем оплаты',
  Unknown = 'Неизвестный статус',
}